+++ lib/wslay_event.c
@@ -113,8 +113,8 @@ static ssize_t wslay_event_frame_send_callback(const uint8_t *data, size_t len,
   return e->ctx->callbacks.send_callback(e->ctx, data, len, e->user_data);
 }
 
-static ssize_t wslay_event_frame_genmask_callback(uint8_t *buf, size_t len,
-                                                  void *user_data)
+static int wslay_event_frame_genmask_callback(uint8_t *buf, size_t len,
+                                              void *user_data)
 {
   struct wslay_event_frame_user_data *e =
     (struct wslay_event_frame_user_data*)user_data;
